Flexible assignation of scheduling pattern (#139)
* Fix evaluation of transmission duration

Change time granularity to picoseconds in the evaluation of the TX 
duration, to avoid wrong scheduling of EndTx.

In the previous version, this led to wrong scheduling when half of the 
symbols are assigned for a transmission and the reamining ones are 
dedicated to another distinct transmission. It never happed before since 
two consecutive slots were never assigned to the same device.

* Fix assignation of consecutive symbols

Since symStart starts from 0, the (+1) is not needed. This problem never 
occured previously since two consecutives slots were never assigned to 
the same device.

* Create flexible scheduling pattern assignation

Define the new method CreateSchedulingPattern, and implemented DEFAULT 
and OPTIMIZED scheduling pattern policies.

Add the possibility of setting the scheduling policy using a class 
attribute.

* Modify examples to run with the new scheduling policy

* Change SchedulingPatternOption to EnumValue

Everything has been updated consequently and the if construct inside the 
method CreateSchedulingPattern has been changed to a switch construct.

* Reduce signal duration by 1 ns

This change has been done to avoid collision on the scheduling of 
consecutive tranmissions in the same same slot.

* Add default option to CreateSchedulingPattern switch